refactor(modules): Complete rename of rocket_launch and main_menu modules
Renames entry files and module classes for the rocket and main modules
to match their folder names, finishing the incomplete work from #161.
- rocket_launch.{cpp,h} -> rocket_module.{cpp,h}, RocketLaunchModule -> RocketModule
- main_menu.{cpp,h} -> main_module.{cpp,h}, MainMenuModule -> MainModule
