+++ src/modules/rocket/contracts_window.cpp
@@ -3,7 +3,7 @@
 #include "launch_window.h"
 #include "modules/engine/gui.h"
 #include "modules/simulation/simulation.h"
-#include "rocket_launch.h"
+#include "rocket_module.h"
 #include <flecs.h>
 #include <spdlog/spdlog.h>
