+++ src/modules/rocket/launch_window.cpp
@@ -5,7 +5,7 @@
 #include "modules/base/base.h"
 #include "modules/engine/gui.h"
 #include "modules/site/site.h"
-#include "rocket_launch.h"
+#include "rocket_module.h"
 #include "widgets/widgets.h"
 #include <algorithm>
 #include <flecs.h>
