+++ src/modules/site/building_window.cpp
@@ -3,10 +3,10 @@
 #include "modules/base/assert.h"
 #include "modules/base/base.h"
 #include "modules/engine/gui.h"
-#include "modules/main/main_menu.h"
+#include "modules/main/main_module.h"
 #include "modules/rocket/actions.h"
 #include "modules/rocket/launch_window.h"
-#include "modules/rocket/rocket_launch.h"
+#include "modules/rocket/rocket_module.h"
 #include "modules/stats/stats.h"
 #include "rocket_prefab_window.h"
 #include "site.h"
